Micro Services/Graph/requirements.txt
gunicorn
flask
stanford_openie
graphviz
